translatable widget utils
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/lib/widget-utils/trunk@1500 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ widget-utils.php
@@ -1,108 +1,134 @@
 <?php
-class widget_utils
-{
-	#
-	# post_meta_boxes()
-	#
+/*
+ * Widget Utils
+ * Author: Denis de Bernardy <http://www.mesoconcepts.com>
+ * Version: 2.0
+ */
+
+/**
+ * widget_utils
+ *
+ * @package Widget Utils
+ **/
+
+class widget_utils {
+	/**
+	 * post_meta_boxes()
+	 *
+	 * @return void
+	 **/
 	
-	function post_meta_boxes()
-	{
+	function post_meta_boxes() {
 		static $done = false;
 		
-		if ( $done ) return;
+		if ( $done )
+			return;
 		
-		add_meta_box('post_widget_config', 'This Post In Widgets', array('widget_utils', 'post_widget_config'), 'post');
+		add_meta_box('post_widget_config', __('This Post In Widgets', widget_utils_textdomain), array('widget_utils', 'post_widget_config'), 'post');
 		add_action('save_post', array('widget_utils', 'post_save_widget_config'));
 
 		$done = true;
 	} # post_meta_boxes()
 	
 	
-	#
-	# page_meta_boxes()
-	#
+	/**
+	 * page_meta_boxes()
+	 *
+	 * @return void
+	 **/
 	
-	function page_meta_boxes()
-	{
+	function page_meta_boxes() {
 		static $done = false;
 		
-		if ( $done ) return;
+		if ( $done )
+			return;
 		
-		add_meta_box('page_widget_config', 'This Page In Widgets', array('widget_utils', 'page_widget_config'), 'page');
+		add_meta_box('page_widget_config', __('This Page In Widgets', widget_utils_textdomain), array('widget_utils', 'page_widget_config'), 'page');
 		add_action('save_post', array('widget_utils', 'page_save_widget_config'));
 		
 		$done = true;
 	} # page_meta_boxes()
 	
 	
-	#
-	# post_widget_config()
-	#
+	/**
+	 * post_widget_config()
+	 *
+	 * @param object $post
+	 * @return void
+	 **/
 	
-	function post_widget_config($post)
-	{
+	function post_widget_config($post) {
 		widget_utils::widget_config('post', $post);
 	} # post_widget_config()
 	
 	
-	#
-	# page_widget_config()
-	#
+	/**
+	 * page_widget_config()
+	 *
+	 * @param object $post
+	 * @return void
+	 **/
 	
-	function page_widget_config($post)
-	{
+	function page_widget_config($post) {
 		widget_utils::widget_config('page', $post);
 	} # page_widget_config()
 	
 	
-	#
-	# post_save_widget_config()
-	#
+	/**
+	 * post_save_widget_config()
+	 *
+	 * @param int $post_ID
+	 * @return void
+	 **/
 	
-	function post_save_widget_config($post_ID)
-	{
-		return widget_utils::save_widget_config($post_ID, 'post');
+	function post_save_widget_config($post_ID) {
+		widget_utils::save_widget_config($post_ID, 'post');
 	} # post_save_widget_config()
 	
 	
-	#
-	# page_save_widget_config()
-	#
+	/**
+	 * page_save_widget_config()
+	 *
+	 * @param int $post_ID
+	 * @return void
+	 **/
 	
-	function page_save_widget_config($post_ID)
-	{
-		return widget_utils::save_widget_config($post_ID, 'page');
+	function page_save_widget_config($post_ID) {
+		widget_utils::save_widget_config($post_ID, 'page');
 	} # page_save_widget_config()
 	
 	
-	#
-	# widget_config()
-	#
+	/**
+	 * widget_config()
+	 *
+	 * @param string $type
+	 * @param object $post
+	 * @return void
+	 **/
 	
-	function widget_config($type, $post)
-	{
+	function widget_config($type, $post) {
 		$post_ID = $post->ID;
 
 		echo '<p>'
-			. 'The following fields let you configure options shared by:'
-			. '</p>';
+			. __('The following fields let you configure options shared by:', widget_utils_textdomain)
+			. '</p>' . "\n";
 
 		echo '<ul>';
 		do_action($type . '_widget_config_affected');
-		echo '</ul>';
+		echo '</ul>' . "\n";
 		
 		echo '<p>'
-			. 'It will <b>NOT</b> affect anything else. In particular WordPress\'s built-in Pages widget. (Use the Silo Pages widget instead.)'
-			. '</p>';
+			. __('It will <b>NOT</b> affect anything else. In particular WordPress\'s built-in Pages widget.', widget_utils_textdomain)
+			. '</p>' . "\n";
 		
 		echo '<table style="width: 100%;">';
 		
 		echo '<tr valign="top">' . "\n"
 			. '<th scope="row" width="120px;">'
-			. 'Title'
+			. __('Title', widget_utils_textdomain)
 			. '</th>' . "\n"
 			. '<td>'
-			. '<input type="text" size="58" style="width: 90%;" tabindex="5"'
+			. '<input type="text" size="58" class="widefat" tabindex="5"'
 			. ' name="widgets_label"'
 			. ' value="' . esc_attr(get_post_meta($post_ID, '_widgets_label', true)) . '"'
 			. ' />'
@@ -111,12 +137,10 @@ class widget_utils
 		
 		echo '<tr valign="top">' . "\n"
 			. '<th scope="row">'
-			. 'Description'
+			. __('Description', widget_utils_textdomain)
 			. '</th>' . "\n"
 			. '<td>'
-			. '<textarea size="58" style="width: 90%;" tabindex="5"'
-			. ' name="widgets_desc"'
-			. ' />'
+			. '<textarea size="58" class="widefat" tabindex="5" name="widgets_desc" />'
 			. format_to_edit(get_post_meta($post_ID, '_widgets_desc', true))
 			. '</textarea>'
 			. '</td>' . "\n"
@@ -124,7 +148,7 @@ class widget_utils
 		
 		echo '<tr valign="top">' . "\n"
 			. '<th scope="row">'
-			. 'Exclude'
+			. __('Exclude', widget_utils_textdomain)
 			. '</th>' . "\n"
 			. '<td>'
 			. '<label>'
@@ -136,7 +160,7 @@ class widget_utils
 				)
 			. ' />'
 			. '&nbsp;'
-			. 'Exclude this entry from automatically generated lists'
+			. __('Exclude this entry from automatically generated lists', widget_utils_textdomain)
 			. '</label>'
 		 	. '</td>' . "\n"
 			. '</tr>' . "\n";
@@ -155,68 +179,57 @@ class widget_utils
 				)
 			. ' />'
 			. '&nbsp;'
-			. '... except for silo stub, silo map and smart links.'
+			. __('... except for silo stubs, silo maps and smart links.', widget_utils_textdomain)
 			. '</label>'
 		 	. '</td>' . "\n"
 			. '</tr>' . "\n";
 		
-		echo '</table>';
+		echo '</table>' . "\n";
 	} # widget_config()
 	
 	
-	#
-	# save_widget_config()
-	#
+	/**
+	 * save_widget_config()
+	 *
+	 * @param int $post_ID
+	 * @param string $type
+	 * @return void
+	 **/
 	
-	function save_widget_config($post_ID, $type = null)
-	{
+	function save_widget_config($post_ID, $type = null) {
 		$post = get_post($post_ID);
 		
-		if ( $post->post_type == 'revision' ) return;
+		if ( $post->post_type == 'revision' || !$_POST || $post->post_type != $type )
+			return;
 		
-		if ( !empty($_POST) )
-		{
-			$post =& get_post($post_ID);
+		if ( isset($_POST['widgets_exclude']) ) {
+			update_post_meta($post_ID, '_widgets_exclude', '1');
 			
-			if ( $post->post_type != $type )
-			{
-				return;
-			}
-		
+			if ( isset($_POST['widgets_exception']) )
+				update_post_meta($post_ID, '_widgets_exception', '1');
+			else
+				delete_post_meta($post_ID, '_widgets_exception');
+		} else {
 			delete_post_meta($post_ID, '_widgets_exclude');
 			delete_post_meta($post_ID, '_widgets_exception');
-			delete_post_meta($post_ID, '_widgets_label');
-			delete_post_meta($post_ID, '_widgets_desc');
-
-			if ( $_POST['widgets_exclude'])
-			{
-				add_post_meta($post_ID, '_widgets_exclude', '1', true);
-				
-				if ( $_POST['widgets_exception'])
-				{
-					add_post_meta($post_ID, '_widgets_exception', '1', true);
-				}
-			}
-			
-			$label = trim(strip_tags(stripslashes($_POST['widgets_label'])));
-			
-			if ( $label )
-			{
-				add_post_meta($post_ID, '_widgets_label', $label, true);
-			}
-			
-			if ( current_user_can('unfiltered_html') )
-				$desc = stripslashes( $_POST['widgets_desc'] );
-			else
-				$desc = stripslashes(wp_filter_post_kses(stripslashes($_POST['widgets_desc'])));
-			
-			if ( $desc )
-			{
-				add_post_meta($post_ID, '_widgets_desc', $desc, true);
-			}
 		}
 		
-		return $post_ID;
+		$label = addslashes(trim(strip_tags(stripslashes($_POST['widgets_label']))));
+		
+		if ( $label )
+			update_post_meta($post_ID, '_widgets_label', $label);
+		else
+			delete_post_meta($post_ID, '_widgets_label');
+		
+		if ( current_user_can('unfiltered_html') )
+			$desc = trim($_POST['widgets_desc']);
+		else
+			$desc = trim(wp_filter_post_kses(stripslashes($_POST['widgets_desc']))));
+		
+		if ( $desc )
+			update_post_meta($post_ID, '_widgets_desc', $desc);
+		else
+			delete_post_meta($post_ID, '_widgets_desc');
 	} # save_widget_config()
 } # widget_utils
 ?>
